refactor: Move code elements to separate files
