+++ design-system/components/code/code.tsx
@@ -5,7 +5,9 @@ import { toJsxRuntime } from 'hast-util-to-jsx-runtime'
 import { Fragment, jsx, jsxs } from 'react/jsx-runtime'
 import { CodeProps, PreElementProps, CodeElementProps } from './code.types'
 import { isTerminal } from './code.helpers'
-import { Pre, Code as CodeElement, PreWrapper } from './code.styles'
+import { Code as CodeElement } from './code.styles'
+import { PreWrapper } from './pre-wrapper'
+import { Pre } from './pre'
 import { CodeHeader } from './code-header'
 import { CodeTitle } from './code-title'
 import { CodeLanguage } from './code-language'
