+++ design-system/components/code/pre-wrapper.tsx
@@ -1,0 +1,4 @@
+import { styled } from '@/styled-system/jsx'
+import { preWrapper } from './pre-wrapper.styles'
+
+export const PreWrapper = styled('div', preWrapper)
