+++ design-system/components/code/pre.styles.tsx
@@ -1,0 +1,11 @@
+import { cva } from '@/styled-system/css'
+
+export const pre = cva({
+  base: {
+    overflowX: 'auto',
+    paddingY: {
+      base: 's',
+      md: 'm'
+    }
+  }
+})
