+++ design-system/components/code/pre.tsx
@@ -1,0 +1,4 @@
+import { styled } from '@/styled-system/jsx'
+import { pre } from './pre.styles'
+
+export const Pre = styled('pre', pre)
